se culmina envio de informes semanales

+++ src/main/java/domain/incidentes/rankings/RepeticionEntidadesComparador.java
@@ -6,17 +6,17 @@ import java.util.Comparator;
 import java.util.Map;
 
 class RepeticionEntidadesComparador implements Comparator<Entidad> {
-    private final Map<Entidad, Integer> conteoEntidades;
+    private final Map<Entidad, Double> conteoEntidades;
 
-    public RepeticionEntidadesComparador(Map<Entidad, Integer> conteoEntidades){
+    public RepeticionEntidadesComparador(Map<Entidad, Double> conteoEntidades){
         this.conteoEntidades = conteoEntidades;
     }
 
     @Override
     public int compare(Entidad entidad1, Entidad entidad2){
-        int rep1 = conteoEntidades.get(entidad1);
-        int rep2 = conteoEntidades.get(entidad2);
+        double rep1 = conteoEntidades.get(entidad1);
+        double rep2 = conteoEntidades.get(entidad2);
 
-        return Integer.compare(rep1, rep2);
+        return Double.compare(rep1, rep2);
     }
 }
